requirements.txt
flask
azure-core
azure-ai-textanalytics
python-dotenv